% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted.calculations.R
\name{achieved.growth}
\alias{achieved.growth}
\title{Calculate achieved growth: max.log.OD minus the inoculated log.OD
  
  @param fitted.well object of class \linkS4class{well} with fitted model info
  @param digits number of significant digits to output
  
  @export}
\usage{
achieved.growth(fitted.well, digits = Inf)
}
\description{
Calculate achieved growth: max.log.OD minus the inoculated log.OD
  
  @param fitted.well object of class \linkS4class{well} with fitted model info
  @param digits number of significant digits to output
  
  @export
}

