% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted.calculations.R
\name{auc}
\alias{auc}
\alias{auc.OD}
\title{Compute Area Under The Curve (AUC)
 
  @param fitted.well object of class \linkS4class{well} with fitted model info
  @param start starting time for AUC calculation
  @param end end time for AUC calculation
  @param digits number of significant digits to output
  @param unlog logic, if TRUE, compute AUC on linear OD scale
  @param constant.added constant added in the \emph{log.OD = log(OD - blank + const)} transformation
  
  @return 
  Returns numeric AUC value computed using the well's \emph{nls} or \emph{loess} model.  
  Returns NA if neither \emph{nls} nor \emph{loess} model is defined for the well.
  
  @seealso \linkS4class{well}
  
  @export}
\usage{
auc(fitted.well, start = NULL, end = NULL, digits = 3, unlog = FALSE,
  constant.added = 0)

auc.OD(fitted.well, start = NULL, end = NULL, digits = 3,
  constant.added = 0)
}
\description{
Compute Area Under The Curve (AUC)
 
  @param fitted.well object of class \linkS4class{well} with fitted model info
  @param start starting time for AUC calculation
  @param end end time for AUC calculation
  @param digits number of significant digits to output
  @param unlog logic, if TRUE, compute AUC on linear OD scale
  @param constant.added constant added in the \emph{log.OD = log(OD - blank + const)} transformation
  
  @return 
  Returns numeric AUC value computed using the well's \emph{nls} or \emph{loess} model.  
  Returns NA if neither \emph{nls} nor \emph{loess} model is defined for the well.
  
  @seealso \linkS4class{well}
  
  @export
}
\section{Functions}{
\itemize{
\item \code{auc.OD}: Compute Area Under The Curve (AUC) on linear OD scale
}}

