% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fit.model.R
\name{fit.nls.model}
\alias{fit.nls.model}
\title{Fit nls model to a well using a specified model}
\usage{
fit.nls.model(input.well, model)
}
\arguments{
\item{input.well}{object of class well}

\item{model}{object of class model, e.g. richards, gompertz or logistic}
}
\description{
Fit nls model to a well using a specified model
}

