% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GCAT.main.R
\name{gcat.analysis.main}
\alias{gcat.analysis.main}
\title{Analyze screening growth data from the given .csv files.}
\usage{
gcat.analysis.main(file.list, single.plate, layout.file = NULL,
  out.dir = getwd(), graphic.dir = paste(out.dir, "/pics", sep = ""),
  add.constant = 0, blank.value, start.index, growth.cutoff = 0.05,
  use.linear.param = F, use.loess = F, smooth.param = 0.1,
  lagRange = NA, totalRange = NA, totalODRange = NA, specRange = NA,
  points.to.remove = 0, remove.jumps = F, time.input = NA,
  plate.nrow = 8, plate.ncol = 12, input.skip.lines = 0,
  multi.column.headers = c("Plate.ID", "Well", "OD", "Time"),
  single.column.headers = c("", "A1"), layout.sheet.headers = c("Strain",
  "Media Definition"), silent = T, verbose = F, return.fit = F,
  overview.jpgs = T)
}
\arguments{
\item{file.list}{A list of full paths to .csv files. all files must be in the same format (see <single.plate>)}

\item{single.plate}{The file in the single plate (wide) format vs. the multi-plate (long) format?}

\item{layout.file}{Full path to a layout file with strain and media definitions (applies to all files in list)}

\item{out.dir}{A directory to output the table of curve parameters to (defaults to working directory)}

\item{graphic.dir}{A directory to output the images of the fitted curves to (defaults to subdirectory "pics" of <out.dir> above)}

\item{add.constant}{A numeric constant that will be added to each curve before the log transform (defaults to 1)}

\item{blank.value}{User can enter a blank OD measurement for uninoculated wells. if NULL, assumes the first OD measurement of each well to be blank.}

\item{start.index}{Which timepoint should be used as the first one after inoculation (defaults to the 2th one)}

\item{growth.cutoff}{Minimum threshold for curve growth.}

\item{use.linear.param}{Whether to use linear parameters or not?}

\item{use.loess}{Whether to use LOESS model or not?}

\item{smooth.param}{Smoothing parameter for LOESS model.}

\item{lagRange}{The heatmap specific range for lag time.}

\item{totalRange}{The heatmap specific range for the achieved growth on log scale.}

\item{totalODRange}{The heatmap specific range for the achieved growth on linear (OD) scale.}

\item{specRange}{The heatmap specific range for spec growth rate.}

\item{points.to.remove}{A list of numbers referring to troublesome points that should be removed across all wells.}

\item{remove.jumps}{Should the slope checking function be on the lookout for large jumps in OD?}

\item{time.input}{The time setting in which the current system is running?}

\item{plate.nrow}{The number of rows in a plate.}

\item{plate.ncol}{The number of columns in a plate.}

\item{input.skip.lines}{If specified, this number of lines shall be skipped from the top when reading the input file with read.csv}

\item{multi.column.headers}{The headers of the result tabular data when analyzing multiple plates at once.}

\item{single.column.headers}{The headers of the result tebaular data when analyzaing a single plate.}

\item{layout.sheet.headers}{The headers of the layout file?}

\item{silent}{Shoulde messages be returned to the console?}

\item{verbose}{Should sub-functions return messages to console? (when I say verbose, I mean it!)}

\item{return.fit}{Whether should a fit well object is returned or not.}

\item{overview.jpgs}{Should GCAT enable an overview image?}
}
\value{
Depending on return.fit setting, an array of fitted well objects or a list of output files
}
\description{
Top-level GCAT function
}

