% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/table2well.R
\name{gcat.append.arrays}
\alias{gcat.append.arrays}
\title{Function to combine two well array datasets by plate}
\usage{
gcat.append.arrays(...)
}
\arguments{
\item{...}{any number of array objects as long as they are all output straight from <load.data>}
}
\description{
This function can append together arrays created using <load.data>
}

