% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gcat.fit.main}
\alias{gcat.fit.main}
\title{Main analysis function for GCAT}
\usage{
gcat.fit.main(file.name, input.data = NULL, load.type = "csv",
  layout.file = NULL, single.plate = F, blank.value = NULL,
  start.index = 2, time.input = NA, normalize.method = "default",
  add.constant = 1, use.log = T, points.to.remove = 0,
  use.linear.param = F, use.loess = F, smooth.param = 0.1,
  fall.cutoff = -0.0025, growth.cutoff = 0.05, remove.jumps = F,
  plate.nrow = 8, plate.ncol = 12, input.skip.lines = 0,
  multi.column.headers = c("Plate.ID", "Well", "OD", "Time"),
  single.column.headers = c("", "A1"), layout.sheet.headers = c("Strain",
  "Media Definition"), growth.model = NA, backup.growth.model = NA,
  silent = F, verbose = F)
}
\arguments{
\item{file.name}{Complete path and file name of a comma-separated values (.csv) file containing growth curve data
in the multiple-plate (long) format.}

\item{input.data}{A list of tables representing input files read with \code{read.table}. Used to save time in cases
of running multiple analyses on the same dataset. If used, the function will ignore \code{file.name} entirely.}

\item{load.type}{.csv by default.}

\item{layout.file}{Specifies the location of a layout file containing identifying information.}

\item{single.plate}{Whether the GCAT is analyzing a single plate or not.}

\item{blank.value}{Blank OD measurement for uninoculated wells. By default(NULL), the value of the first OD
measurement in each well is used.}

\item{start.index}{Which timepoint should be used as the first one after inoculation?}

\item{time.input}{Either a character describing the format used to convert timestamps in the input to numbers
representing number of seconds (see \code{strptime}), or a factor to divide entries in the Time column by to get the
numbers of hours.}

\item{normalize.method}{Describes the method used by \code{normalize.ODs} to normalize cell density values using blank reads.}

\item{add.constant}{A value for r in the log(OD + r) transformation.}

\item{use.log}{Should the analysis use log on all values.}

\item{points.to.remove}{A vector of integers specifying which timepoints should be removed across all wells.
By default(0) none are marked for removal.}

\item{use.linear.param}{Should the linear parameter be used or not.}

\item{use.loess}{Should the loess model be used or not.}

\item{smooth.param}{If loess model is used, this parameter define the smoothing parameter for the loess model.}

\item{fall.cutoff}{A cutoff used by \code{check.slopes} to decide on thresholds for jumps and tanking.}

\item{growth.cutoff}{A threshold used by check.growth to decide whether a well displays growth.}

\item{remove.jumps}{Should jumps in OD detected by the subfunction \code{check.slopes}?}

\item{plate.nrow}{The number of rows in the input files.}

\item{plate.ncol}{The number of columns in the input files.}

\item{input.skip.lines}{If specified, this number of lines shall be skipped from the top when reading the input file with read.csv}

\item{multi.column.headers}{The headers of the column when analyzing multiple plates.}

\item{single.column.headers}{The headers of the column when analyzing a single plate.}

\item{growth.model}{What growth model should be used?}

\item{backup.growth.model}{If the main growth model fails, the back up model will be used.}

\item{silent}{Surpress all messages.}

\item{verbose}{Display all messages when analyzing each well.}

\item{layour.sheet.headers}{The headers of the layout file.}
}
\value{
An array of well objects
}
\description{
This is the main function that handles all the analyses for data files in both single and multiple plate formats.
It is called by the top level function \code{gcat.analysis.main} along with \code{gcat.output.main}.
}

