% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gcat.output.main}
\alias{gcat.output.main}
\title{Output function for generating files from fitted data.}
\usage{
gcat.output.main(fitted.well.array, out.prefix = "", source.file.list,
  upload.timestamp = NULL, add.constant, blank.value, start.index,
  growth.cutoff, points.to.remove, remove.jumps, out.dir = getwd(),
  graphic.dir = paste(out.dir, "/pics", sep = ""), overview.jpgs = T,
  use.linear.param = F, use.loess = F, plate.nrow = 8, plate.ncol = 12,
  unlog = F, silent = T)
}
\arguments{
\item{fitted.well.array}{A list of fitted well objects.}

\item{out.prefix}{Prefix that is in the name of output files.}

\item{blank.value}{User can enter a blank OD measurement for uninoculated wells.
If NULL, defaults to the value of the first OD measurement of each well.}

\item{start.index}{Which timepoint should be used as the first one after inoculation (defaults to the 2th one)}

\item{growth.cutoff}{Minimum threshold for curve growth.}

\item{points.to.remove}{A list of numbers referring to troublesome points that should be removed across all wells.}

\item{remove.jumps}{Should the slope checking function be on the lookout for large jumps in OD?}

\item{out.dir}{name a directory to output the table of curve parameters to (defaults to working directory)}

\item{graphic.dir}{name a directory to output the images of the fitted curves to
(defaults to subdirectory "pics" of <out.dir> above)}

\item{overview.jpgs}{should jpgs be generated for each plate with the overview graphic?
This is for backwards compatibility with the old web server.}

\item{unlog}{should exported graphics be transformed back to the OD scale?}

\item{silent}{should messages be returned to the console?}

\item{constant.added}{(should be the same value as add.constant above) -
used to readjust for the constant added during the log transform when plotting ODs.}
}
\value{
A list of output files if success.
}
\description{
Handles files and directories, calls \code{table.out}, \code{plate.overview} and \code{view.fit}
to generate output tables and graphics.
}

