% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/table2well.R
\name{gcat.start.times}
\alias{gcat.start.times}
\title{Convert timestamps to hours from start and sort timepoints
This function acts on a single well and modifies the raw data stored in slot "screen.data"}
\usage{
gcat.start.times(input.well, time.input, start.index = 2)
}
\arguments{
\item{input.well}{an object of class well}

\item{time.input}{specifies time format in the input. allowed values are "%S", for seconds, "%d", for days, or anything complying with
ISO C / POSIX standards; see <strptime>.}

\item{start.index}{which timepoint should be used as the starting time at inoculation?}
}
\description{
Convert timestamps to hours from start and sort timepoints
This function acts on a single well and modifies the raw data stored in slot "screen.data"
}
\details{
note: reading directly from excel to R results in timestamps being converted to days.
}

