% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fitted.calculations.R
\name{inoc.log.OD}
\alias{inoc.log.OD}
\title{Calculate log.OD at inoculation from fitted parameters}
\usage{
inoc.log.OD(fitted.well, digits = Inf)
}
\description{
Calculate log.OD at inoculation from fitted parameters
}

