% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/class.well.R
\docType{methods}
\name{length,well-method}
\alias{length,well-method}
\title{Get the number of data points in a well}
\usage{
\S4method{length}{well}(x)
}
\arguments{
\item{x}{object of class \code{well}}
}
\description{
Get the number of data points in a well
}

