% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/normalize.and.transform.R
\name{normalize.ODs}
\alias{normalize.ODs}
\title{Normalize OD readings for an entire array of well objects}
\usage{
normalize.ODs(well.array, normalize.method = "default", blank.value = NULL,
  start.index = 2, add.constant = 1)
}
\arguments{
\item{well.array}{an array of well objects.}

\item{normalize.method}{see Details}

\item{blank.value}{user can enter a blank OD measurement for uninoculated wells. if NULL, defaults to the value of the first OD measurement of each well.}

\item{start.index}{which timepoint should be used as the first one after inoculation (defaults to the 2th one)}

\item{add.constant}{add a numeric constant to all timepoints in all wells.}
}
\description{
Normalize OD readings for an entire array of well objects
}
\details{
Note: This function does not write any new OD values to the well objects in the array - it only
  fills the "norm" slot of each well object in the array with a value that will be subtracted
  from all OD measurements when returning data from the wells using the function <data.from> (see well.class.R)

 These functions make use of <raw.data> which simply returns the raw time and OD of a well (also see well.class.R)

 note this is the only normalization function that acts on an entire array instead of an individual well.

 normalize.method settings:
 \describe{
 \item{default}{subtracts the blank OD (either specified by <blank.value> or taken from the first timepoint as default)
 of each well from all timepoints in that well}
 \item{average.blank}{subtracts the mean of all first OD timepoints on a plate from all timepoints in all wells on that plate}
 \item{average.first}{takes the mean of the difference between the OD of the specified <start> timepoint and the first timepoint of all wells on a plate
                    and subtracts this value from all timepoints in all wells on that plate}
 \item{anything else}{do nothing}
 }
}

