% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/class.well.R
\docType{methods}
\name{plot,well,missing-method}
\alias{plot,well,missing-method}
\title{A plot method for well}
\usage{
\S4method{plot}{well,missing}(x, y, constant.added = 1, xlim = NULL,
  ylim = NULL, well.number = NULL, scale = 1, number.points = T,
  draw.symbols = F, show.text = T, show.calc = T, draw.guess = NULL,
  auc.start = NULL, auc.end = NULL, ...)
}
\arguments{
\item{x}{object of class well}

\item{y}{not used}

\item{constant.added}{used to readjust for the constant added during the log transform: log.OD = log(OD - blank + constant.added)}

\item{xlim}{x axis limits, vector of length 2}

\item{ylim}{y axis limits, vector of length 2}

\item{well.number}{the number of the well in an array of wells}

\item{scale}{determines the font scale for the entire graph. all cex values are calculated from this}

\item{number.points}{should points be labeled with numeric indices?}

\item{draw.symbols}{should <check.slopes> be called on the well and markings drawn on the graph?}

\item{show.text}{show R^2 and growth curve parameters as text on the plot}

\item{show.calc}{draw lines that illustrate growth curve parameters}

\item{draw.guess}{initial guess model.  Drawn if specified}

\item{...}{additional arguments passed to the generic plot function}
}
\description{
A plot method for well
}

