% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.fit.R
\name{plot_data}
\alias{plot_data}
\title{plot_data}
\usage{
plot_data(input.well, view.raw.data = F, unlog = F, scale = 1,
  main = paste(plate.name(input.well), well.name(input.well)),
  number.points = T, draw.symbols = F, constant.added, ylim, ...)
}
\arguments{
\item{input.well}{The well object that need to be plottedd}

\item{view.raw.data}{should the raw data be plotted? (}

\item{unlog}{should data be plotted on a linear (vs. logarithmic) scale?}

\item{scale}{determines the font scale for the entire graph. all cex values are calculated from this.}

\item{main}{...}

\item{number.points}{should points be labeled with numeric indices?}

\item{draw.symbols}{- should <check.slopes> be called on the well and markings drawn on the graph?}

\item{constant.added}{Similar to added.constant.}

\item{ylim}{...}

\item{...}{additional arguments passed to plot()}
}
\description{
Basic function plots time vs. OD from a well object
}

