% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/table.output.R
\name{table.out}
\alias{table.out}
\title{Populate an output table with parameters and other useful info for each well in a fitted dataset.}
\usage{
table.out(fitted.data.set, unlog = F, constant.added, reach.cutoff = 0.9,
  filename.timestamp = NULL, use.linear.param = F, use.loess = F,
  auc.start = NULL, auc.end = NULL)
}
\arguments{
\item{fitted.data.set}{array of fitted well objects}

\item{unlog}{- Should OD values be returned on the linear scale instead of log-transformed scale?}

\item{constant.added}{- For returning values on linear scale, what constant was added to ODs before the log transform?}

\item{reach.cutoff}{- what proportion of the plateau OD must tbe reached by the last valid timepoint for the curve to be marked as reaching
its plateau OD?}

\item{filename.timestamp}{timestamp for addition to output file names (for file references in last column of the output)}

\item{use.linear.param}{did the model formula contain a linear parameter?  Should normally be FALSE, as the linear parameter is deprecated}

\item{use.loess}{was LOESS used to fit the data (instead of unsing a growth curve model formula)?}

\item{auc.start}{Start time for AUC computation}

\item{auc.end}{End time for AUC computation}
}
\description{
Populate an output table with parameters and other useful info for each well in a fitted dataset.
}

