% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/class.well.R
\docType{methods}
\name{well-accessors}
\alias{getAddInfo}
\alias{getAddInfo,well-method}
\alias{getCurPar}
\alias{getCurPar,well-method}
\alias{getEquation}
\alias{getEquation,well-method}
\alias{getFitErr}
\alias{getFitErr,well-method}
\alias{getFitInfo}
\alias{getFitInfo,well-method}
\alias{getFitPar}
\alias{getFitPar,well-method}
\alias{getInflectionTime}
\alias{getInflectionTime,well-method}
\alias{getLoess}
\alias{getLoess,well-method}
\alias{getModelName}
\alias{getModelName,well-method}
\alias{getNorm}
\alias{getNorm,well-method}
\alias{getPosition}
\alias{getPosition,well-method}
\alias{getRSS}
\alias{getRSS,well-method}
\alias{getScreenData}
\alias{getScreenData,well-method}
\alias{getStartIndex}
\alias{getStartIndex,well-method}
\alias{getUseLog}
\alias{getUseLog,well-method}
\alias{getWellInfo}
\alias{getWellInfo,well-method}
\alias{getnls}
\alias{getnls,well-method}
\alias{well-accessors}
\title{Accessors for the well class}
\usage{
getPosition(object)

\S4method{getPosition}{well}(object)

getWellInfo(object)

\S4method{getWellInfo}{well}(object)

getScreenData(object)

\S4method{getScreenData}{well}(object)

getStartIndex(object)

\S4method{getStartIndex}{well}(object)

getUseLog(object)

\S4method{getUseLog}{well}(object)

getNorm(object)

\S4method{getNorm}{well}(object)

getCurPar(object)

\S4method{getCurPar}{well}(object)

getFitErr(object)

\S4method{getFitErr}{well}(object)

getEquation(object)

\S4method{getEquation}{well}(object)

getModelName(object)

\S4method{getModelName}{well}(object)

getFitInfo(object)

\S4method{getFitInfo}{well}(object)

getAddInfo(object)

\S4method{getAddInfo}{well}(object)

getInflectionTime(object)

\S4method{getInflectionTime}{well}(object)

getRSS(object)

\S4method{getRSS}{well}(object)

getLoess(object)

\S4method{getLoess}{well}(object)

getnls(object)

\S4method{getnls}{well}(object)

getFitPar(object)

\S4method{getFitPar}{well}(object)
}
\arguments{
\item{object}{object of class \linkS4class{well}}
}
\description{
Accessors for the well class
}

