% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fitted.calculations.R
\name{well.eval}
\alias{well.eval}
\title{Evaluate estimated OD at any timepoints using the fitted model}
\usage{
well.eval(fitted.well, Time = NULL)
}
\description{
Evaluate estimated OD at any timepoints using the fitted model
}

