% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fitted.calculations.R
\name{auc}
\alias{auc}
\title{Compute Area Under The Curve (AUC)}
\usage{
auc(fitted.well, start = NULL, end = NULL, digits = 3)
}
\arguments{
\item{fitted.well}{object of class \linkS4class{well} with fitted model info}

\item{start}{starting time for AUC calculation}

\item{end}{end time for AUC calculation}

\item{digits}{number of significant digits to output}
}
\value{
Returns numeric AUC value computed using the well's nls or loess model.
  Returns NA if neither nls nor loess model is defined for the well.
}
\description{
Compute Area Under The Curve (AUC)
}

