% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.fit.R
\name{create.heatmap}
\alias{create.heatmap}
\title{Create a heat map of a plate}
\usage{
create.heatmap(fitted.well.array, attribute, MinMax = NA, constant.added)
}
\arguments{
\item{fitted.well.array}{matrix containing well array object data}

\item{attribute}{the data type we should use to create a heatmap}

\item{MinMax}{The specific range for the heatmap.}

\item{constant.added}{the numeric constant that was added to each curve before the log transform:
same as \code{add.constant} in \link{gcat.analysis.main}}
}
\value{
path of heatmap pdf file
}
\description{
Create a heat map of a plate
}
\details{
This function is used to create a heatmap using
specific growth, total growth, or lag time
for each well on a plate.
}

