% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GCAT.main.R
\name{gcat.output.main}
\alias{gcat.output.main}
\title{Output function for generating files from fitted data.}
\usage{
gcat.output.main(fitted.well.array, out.prefix = "", source.file.list,
  upload.timestamp = NULL, add.constant, blank.value, start.index,
  growth.cutoff, points.to.remove, remove.jumps, out.dir = getwd(),
  graphic.dir = paste(out.dir, "/pics", sep = ""), overview.jpgs = T,
  use.linear.param = F, use.loess = F, lagRange = NA, totalRange = NA,
  totalODRange = NA, specRange = NA, plate.nrow = 8, plate.ncol = 12,
  unlog = F, silent = T, main.envir)
}
\arguments{
\item{fitted.well.array}{A list of fitted well objects.}

\item{out.prefix}{Prefix that is in the name of output files.}

\item{source.file.list}{A list of the source files' names.}

\item{upload.timestamp}{The time format indicated by the user.}

\item{add.constant}{used to readjust for the constant added during the log transform when plotting ODs.}

\item{blank.value}{User can enter a blank OD measurement for uninoculated wells.
If NULL, defaults to the value of the first OD measurement of each well.}

\item{start.index}{Which timepoint should be used as the first one after inoculation (defaults to the 2th one)}

\item{growth.cutoff}{Minimum threshold for curve growth.}

\item{points.to.remove}{A list of numbers referring to troublesome points that should be removed across all wells.}

\item{remove.jumps}{Should the slope checking function be on the lookout for large jumps in OD?}

\item{out.dir}{name a directory to output the table of curve parameters to (defaults to working directory)}

\item{graphic.dir}{name a directory to output the images of the fitted curves to
(defaults to subdirectory "pics" of <out.dir> above)}

\item{overview.jpgs}{should jpgs be generated for each plate with the overview graphic?
This is for backwards compatibility with the old web server.}

\item{use.linear.param}{linear parameter is used or not?}

\item{use.loess}{Is LOESS model going to be used?}

\item{lagRange}{The heatmap specific range for lag time.}

\item{totalRange}{The heatmap specific range for the achieved growth.}

\item{totalODRange}{The heatmap specific range for the achieved growth on linear (OD) scale.}

\item{specRange}{The heatmap specific range for spec growth rate.}

\item{plate.nrow}{The number of rows for a plate.}

\item{plate.ncol}{The number of columns for a plate}

\item{unlog}{should exported graphics be transformed back to the OD scale?}

\item{silent}{should messages be returned to the console?}

\item{main.envir}{starting environment of gcat.analysis.main(), captured as a list, printed out for debugging}
}
\value{
A list of output files if success.
}
\description{
Handles files and directories, calls \code{table.out}, \code{plate.overview} and \code{view.fit}
to generate output tables and graphics.
}

