% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/class.model.R
\name{model}
\alias{model}
\title{Model}
\usage{
model(name, expression, formula, guess)
}
\arguments{
\item{name}{The name of the model}

\item{expression}{Expression of the model}

\item{formula}{The formula of this model}

\item{guess}{The guess of this model}
}
\value{
The new model
}
\description{
Function to create a new model
}

