% Generated by roxygen2 (4.0.2): do not edit by hand
\name{transform.ODs}
\alias{transform.ODs}
\title{Transform.Ods}
\usage{
\method{transform}{ODs}(input.well, use.log = T, blank.value = NULL,
  start.index = 2, negative.OD.cutoff = 10, constant.added = 1, ...)
}
\arguments{
\item{input.well}{an object of class well}

\item{use.log}{gets added to the "use.log" slot of the well object. this will determine whether the log-transformed data
or raw normalized data is returned using the function \code{data.from}.}

\item{blank.value}{user can enter a blank OD measurement for uninoculated wells. if NULL, defaults to the value of the first OD measurement of each well.}

\item{start.index}{which timepoint should be used as the first one after inoculation (defaults to the 2th one)}

\item{negative.OD.cutoff}{if any ODs below the specified blank value are detected before this index timepoint, the entire well is discarded.}
}
\description{
This function adds a "log.OD" column to the "screen.data" slot of a well object with log-transformed data.
The raw data is kept intact.
It also checks to see if any of the raw OD values (before a certain timepoint) is below the blank OD.
This can be disastrous for the log(OD) transform.
}

