% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/class.well.R
\docType{class}
\name{well-class}
\alias{well-class}
\title{Class that contains well data}
\description{
Class that contains well data
}
\section{Slots}{

\describe{
\item{\code{position}}{- 3-member vector containing identifying information for the well: row (letters), column (numbers) and plate ID.}

\item{\code{well.info}}{- a list containing strain and media names if provided}

\item{\code{screen.data}}{- a data frame with Time and raw OD values. This is the only slot that is filled upon creation of a well object.
               as different functions are run on the well the data frame gets filled with additional columns.}

\item{\code{start.index}}{- integer index of the time point where growth curve starts, e.g. of the inoculation time point}

\item{\code{use.log}}{- a single logical value denoting whether to return log-transformed values when data is requested from the well}

\item{\code{norm}}{- a value to subtract from all OD values before returning data. filled by <normalize.ODs> (see normalize.and.transform.R)}

\item{\code{curve.par}}{- a list of parameters that denote whether the well is empty, whether it contains ODs indicating a viable culture, whether it tanks at a certain timepoint.}

\item{\code{fit.par}}{- will be a list containing the fitted model parameters}

\item{\code{fit.std.err}}{- will be a list containing the standard errors for the fitted model parameters}

\item{\code{equation}}{- will contain an expression for evaluating the successfully fitted model}

\item{\code{model.name}}{- will contain the name of the successfully fit model}

\item{\code{fit.info}}{- a message with info about whether the fit was successful, failed, or skipped.}

\item{\code{add.info}}{- a message with info about whether jumps in OD were detected or removed, or if ODs were detected below the blank OD.}

\item{\code{inflection.time}}{- the Time value at the point where the specific growth is located. no longer a formula param NWD}

\item{\code{rss}}{- residual sum of squares}

\item{\code{loess}}{- object returned by running loess on the normalized well data}

\item{\code{nls}}{- object returned by running nls on the normalized well data}
}}

